C++ buffer to Python list
Jumped the hoops of snek's awful object model in the form of an array containing the buffers allocated to perform the call; the caller can request to get these back so that overwriting of memory can actually be done. But for crying out loud, Python should've just been implemented in a less obtuse way and then low level operations like these wouldn't be so downright inconvenient. Programming for evrrybody am I rite?