CVYC labels
Clairvoyance (CVYC) is AR/speak for the wierd, quasi-mystical feeling of one symbol being seemingly able to interpret or interact with a another that is defined much later, before it is even declared. As applied to labels, this means we are able to define offsets within a segment before we know how big the segment currently is.