Implement cmamclip()
Quality of life feature for writing C code
inside of a CMAM macro. It works as such:

`C [expr]? [{blk}]?;`

The entire expression plus the block that follows
will be then stored as a string. A call to cmamclip()
then returns that string, and textual replacement of
tokens can be achieved like so:

```CMAM
macro {
  C T#_[name] [expr];
  my $code=cmamclip(T=>'value');
};
```

All in all, this makes it much easier to perform
common metaprogramming tasks.