Fix parsing of CMAM macros
The problem is that macros are written in perl,
but we're using Tree::C as parser. I've patched it
up so that macros can be interpreted as strings,
akin to what we did with the 'C' blocks last time.

However, this is a temporal solution. Long-term,
I would like to just write everything in a C-like
syntax (or as C-like as possible). That doesn't mean
ditching perl, but it does mean using a *different*
king of perl ;>